Slimdown: unify environment resolver paths
