+++ docs/PRD-slimdown.md
@@ -14,7 +14,7 @@
 - Adapter build/test definitions repeat identical evaluation defaults (fixture cwd + clean fixture enforcement).
 - Static file serving logic is duplicated between status UI and mirror tooling, including content-type maps and path safety checks.
 - The diagnostics pipeline is defined twice (config + code fallback), increasing drift risk when stages change.
-- Repo/run/out path resolution and small helper logic (date math, path normalization) are reimplemented across orchestrator, design, and scripts with slightly different env fallbacks.
+- Repo/run/out path resolution plus directory listing/date/path helpers are reimplemented across orchestrator, design, and scripts with slightly different env fallbacks.
 - Docs archive/freshness/review tooling and mirror fetch still hardcode `.runs`/`out` roots, ignoring `CODEX_ORCHESTRATOR_*` envs.
 - Status UI dataset build logic duplicates task key normalization behavior already shared elsewhere.
 
@@ -31,7 +31,7 @@
 - Reuse guardrail stage sets for delegation/spec-guard across pipelines to avoid repeated command blocks.
 - Reuse a configurable slugify helper across design pipeline and orchestrator tooling.
 - Consolidate adapter evaluation defaults (fixture cwd + clean fixture enforcement) into shared helpers.
-- Consolidate static file serving helpers and repo/run/out path resolution to avoid drift, including remaining script-level helper duplicates.
+- Consolidate static file serving and repo/run/out path + directory listing helpers to avoid drift, including remaining script-level helper duplicates.
 - Reduce pipeline duplication by consolidating docs-review checks and removing redundant fallback pipelines.
 - Reuse shared task-key normalization in status UI dataset generation to remove drift.
 - Align docs archive/freshness/run-review/mirror outputs with shared repo/run/out resolvers to honor configured env roots.
