+++ scripts/lib/run-manifests.d.ts
@@ -1,6 +1,12 @@
 export function resolveRepoRoot(): string;
 export function resolveRunsDir(repoRoot: string): string;
 export function resolveOutDir(repoRoot: string): string;
+export function resolveEnvironmentPaths(): {
+  repoRoot: string;
+  runsRoot: string;
+  outRoot: string;
+  taskId: string;
+};
 export function listDirectories(dirPath: string): Promise<string[]>;
 export function parseRunIdTimestamp(runId: string): Date | null;
 export function pickLatestRunId(runIds: string[]): string | null;
